requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
mysqlclient==2.2.0
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
pandas==2.0.3
openpyxl==3.1.2
Pillow==10.0.0
Werkzeug==2.3.7
dlib==19.24.2
python-decouple==3.8